DX11::ShaderCompiler and more
- Removed ::Algorithms class
- Moved FixedArray to base namespace
- Removed EffectAsset

+++ 3DRadSpace/Engine3DRadSpace/Graphics/IShader.hpp
@@ -22,14 +22,14 @@ namespace Engine3DRadSpace::Graphics
 		///<summary>
 		/// Sets an entire uniform buffer to the shader.
 		/// </summary>
-		virtual void SetData(unsigned index, const void *data, unsigned dataSize) = 0;
+		virtual void SetData(unsigned index, const void *data, size_t dataSize) = 0;
 		virtual void SetTexture(unsigned index, ITexture2D *texture) = 0;
 		virtual void SetTextures(std::span<ITexture2D*> textures) = 0;
 		virtual void SetSampler(unsigned index, ISamplerState *samplerState) = 0;
 		virtual void SetShader() = 0;
 
 		virtual std::vector<Reflection::IReflectedField*> GetVariables() const = 0;
-		virtual void Set(const std::string& name, const void* data, unsigned dataSize) = 0;
+		virtual void Set(const std::string& name, const void* data, size_t dataSize) = 0;
 
 		template<typename T>
 		void Set(const std::string& name, const T& data)
